requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
alpha-vantage==2.3.1
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.32

